Change directory management
Instead of trying to run this from wherever (a use case that doesn't actually exist), simply move into the directory where the script is located.

This is to deal with a bug where the caption text wasn't being posted (tailing caption.srt was missing the SCRIPT_DIR prefix), but it seems to make more sense to eliminate this brittleness.