Fix how caption text is extracted
This bit is where we extract the caption text from the SRT file, omitting the timestamp ranges and the caption number. As it was written before, it was too simplistic, resulting in some lines of captions being skipped. So instead we get really specific about the two types of lines that get skipped (bare integers and timestamp ranges).